Begin to implement a recursive descent parser.
As yet, it doesn't recognise much syntax yet, but it should serve as a
good example of how easy it is to get stuff to work.

It can recognise labels (in the "marker for a goto" sense of the word)
lines may begin with a label, expression, or a statement - which at this
point means either IF, or PRINT.
it can handle conditions with either a simple expression or two
expressions and a comparison operator.


expressions are currently limited to nothing more than labels, in
order to avoid having to write a lot more code at this point in time.