Parser: Implement expression evaluation.
This implements a "Shunting Yard" type expression evaluator.

It can handle + - * / and ()

This is also the first use of the struct line_entry data by
the parser - it uses it to denote that a given token is a unary operator.

Also of note is that unary operators have been promoted up the precedence
hierarchy, thus fixing a heretofore un-noticed parser bug.