Parser: Clean up unary operator handling
Rather than use a flag in the operators line entry to indicate that it
is unary, instead, allow the expression parser to promote the line entry's
token to a unary operator on the fly.

This has a knock-on effect of cleaning up the rest of the code to handle
unary operators considerably.