Expression: Fix bug introduced when implementing PROC etc.
The AST is subtly different depending on if FN is emitted by the
expression parser code or by the FN call code in the parser.

This is because the expression parser runs in its own AST context.

Its a little bit of a pain to fix, so we'll work around it for now.